Issue #225: Enable useTags for OpenApi specs
